requirements.txt
PyYAML
pdfminer
lxml==3.2.1
wsgiref==0.1.2
